move test directory outside of src directory
