+++ test/stats.ts
@@ -5,7 +5,7 @@ process.env.NODE_ENV = 'test'
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const uuidv1 = require('uuid')
-import { AppServer } from '../index'
+import { AppServer } from '../src/index'
 const expect = chai.expect
 chai.use(chaiHttp)
