+++ test/store.ts
@@ -4,45 +4,47 @@ process.env.NODE_ENV = 'test'
 // Require the dev-dependencies
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-import { AppServer } from '../index'
+import { AppServer } from '../src/index'
 const expect = chai.expect
 chai.use(chaiHttp)
-import { AppConfig } from '../config/config'
 
 // Our parent block
-describe('Swagger', () => {
-  it('it should 200', (done) => {
+describe('Shop', () => {
+  it('it should get store info', (done) => {
     chai.request(AppServer)
-      .get(AppConfig.static_uri + '/swagger.json')
+      .get('/store')
       .end((err, res) => {
         if (err) {
           console.log(err)
         }
         expect(res).to.have.status(200)
+        expect(res.body).to.be.a('object')
         done()
       })
   })
 
-  it('it should 200', (done) => {
+  it('it should get store info in french', (done) => {
     chai.request(AppServer)
-      .get(AppConfig.static_uri + '/swagger.yaml')
+      .get('/store/fr')
       .end((err, res) => {
         if (err) {
           console.log(err)
         }
         expect(res).to.have.status(200)
+        expect(res.body).to.be.a('object')
         done()
       })
   })
 
-  it('it should 200', (done) => {
+  it('it should get store info in it', (done) => {
     chai.request(AppServer)
-      .get('/api-docs')
+      .get('/store/it')
       .end((err, res) => {
         if (err) {
           console.log(err)
         }
         expect(res).to.have.status(200)
+        expect(res.body).to.be.a('object')
         done()
       })
   })
