+++ test/swagger.ts
@@ -4,47 +4,45 @@ process.env.NODE_ENV = 'test'
 // Require the dev-dependencies
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-import { AppServer } from '../index'
+import { AppServer } from '../src/index'
 const expect = chai.expect
 chai.use(chaiHttp)
+import { AppConfig } from '../src/config/config'
 
 // Our parent block
-describe('Shop', () => {
-  it('it should get store info', (done) => {
+describe('Swagger', () => {
+  it('it should 200', (done) => {
     chai.request(AppServer)
-      .get('/store')
+      .get(AppConfig.static_uri + '/swagger.json')
       .end((err, res) => {
         if (err) {
           console.log(err)
         }
         expect(res).to.have.status(200)
-        expect(res.body).to.be.a('object')
         done()
       })
   })
 
-  it('it should get store info in french', (done) => {
+  it('it should 200', (done) => {
     chai.request(AppServer)
-      .get('/store/fr')
+      .get(AppConfig.static_uri + '/swagger.yaml')
       .end((err, res) => {
         if (err) {
           console.log(err)
         }
         expect(res).to.have.status(200)
-        expect(res.body).to.be.a('object')
         done()
       })
   })
 
-  it('it should get store info in it', (done) => {
+  it('it should 200', (done) => {
     chai.request(AppServer)
-      .get('/store/it')
+      .get('/api-docs')
       .end((err, res) => {
         if (err) {
           console.log(err)
         }
         expect(res).to.have.status(200)
-        expect(res.body).to.be.a('object')
         done()
       })
   })
